fix(core,runtime): M4a registry id stability, M11 recycled-slot refs
Both were confirmed bugs kept alive as should_fail tests pending a design
call. Neither decorator remains.

M4a — ComponentRegistry ids are positions in the name-sorted list, and that
sort is what makes them reproducible across builds, since static-init order
across translation units is unspecified. A Register arriving after an id had
been issued re-sorted and renumbered, silently mis-mapping ids that
ComponentIdOf<T> memoises in a function-local static and that saved scenes
store. Decoupling id from sort order (one of the options on file) would trade
the bug for non-determinism, and appending with a fresh id does not work
either: ById()/All() hand out pointers into _metas and _serializable holds
them, so growing the vector dangles them all. Register now refuses once
finalized — error naming the component, then an assert. Ordering matters
there: the assert does not return, so the log has to come first to be seen,
and the assert message is a literal that cannot name the component.

The Core tests registered fixtures lazily inside test cases, which only worked
while late registration silently renumbered; they now register from a static
initializer, before main, exactly as generated registrations do.

M11 — under ScopedRawEntityContext, EntityToIndex returned the bare slot and
IndexToEntity returned that slot's current occupant, so a ref to a destroyed
entity resolved to whoever reused the slot. Rejecting non-IsAlive refs (the
other option on file) does not help: a recycled slot is perfectly alive. The
raw context now packs (slot, generation) through the existing EntityKey and
resolves only on an exact generation match, yielding NullEntity otherwise. The
paired restore revives entities at their original handle, so the intended flow
still resolves exactly. EntityRef codegen widened to 64-bit accordingly; the
normal save path still emits a small dense index, which fits either way.

Validated: 8/8 suites, 85/85 Core cases, zero warnings.

+++ modules/Runtime/include/Assisi/Runtime/SceneSerializer.hpp
@@ -72,13 +72,13 @@ class SceneSerializer
     ///
     /// Only valid to call from within a component's serialize lambda.
     /// Returns nullopt if the entity is unknown or no save is in progress.
-    static std::optional<uint32_t> EntityToIndex(ECS::Entity entity);
+    static std::optional<uint64_t> EntityToIndex(ECS::Entity entity);
 
     /// @brief Map a serial index to the live entity created during the current Load.
     ///
     /// Only valid to call from within a component's addToScene lambda.
     /// Returns NullEntity if the index is out of range or no load is in progress.
-    static ECS::Entity IndexToEntity(uint32_t index);
+    static ECS::Entity IndexToEntity(uint64_t index);
 
     /// @brief RAII scope that makes EntityRef fields serialize/deserialize against
     /// *raw entity handles* instead of remapped serial indices.
@@ -91,8 +91,10 @@ class SceneSerializer
     /// the hierarchy with no error.
     ///
     /// Within this scope the mapping is identity-preserving instead:
-    ///   - EntityToIndex(e) returns `e.index` (the raw slot, no remap);
-    ///   - IndexToEntity(i) returns `scene.EntityAt(i)` (the live handle at that slot).
+    ///   - EntityToIndex(e) returns a packed (slot, generation) key, no remap;
+    ///   - IndexToEntity(k) returns the live handle at that slot, but only when its
+///     generation still matches the one packed into `k` — a recycled slot
+///     resolves to NullEntity rather than to its new occupant.
     /// This is exact only because the paired restore uses Scene::ReviveAt to bring
     /// entities back at their original (index, generation) — EntityAt(index) then
     /// resolves to the same handle that was captured.
